requirements.txt
absl-py==2.1.0
accelerate==1.5.2
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.13.4
biopython==1.85
blinker==1.9.0
botocore==1.38.27
cachetools==6.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==3.4.1
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
distro==1.9.0
einops==0.8.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.64.1
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain-core==0.3.65
langchain-deepseek==0.1.3
langchain-huggingface==0.3.0
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lightgbm==4.4.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.43.1
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.88.0
opencv-python==4.10.0.84
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.3.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spyder-kernels==2.5.2
stack-data==0.6.3
streamlit==1.46.0
sympy==1.13.2
tenacity==9.1.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.49.0
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.17.2
wurlitzer==3.1.1
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
